Update file index for Win32

+++ app/index.php
@@ -37,8 +37,8 @@ header("Content-Security-Policy: default-src 'self'; style-src 'self' 'unsafe-in
     foreach ($files as $file) {
         if (!empty($file)) {
             if ($file !== '.' && $file !== '..') {
-                if (file_exists($file) && $file !== 'index.php' && $file !== 'style.css') {
-                    if (is_file($file)) {
+                if (file_exists(dirname(__FILE__) . '/' . $file) && $file !== 'index.php' && $file !== 'style.css' && $file !== 'router.php') {
+                    if (is_file(dirname(__FILE__) . '/' . $file)) {
                         printf('<li><a href="%s" target="_blank" rel="noopener">%s</a></li>', $file, ucfirst(str_replace('.php', '', $file)));
                     }
                 }
